integrations/openai/pyproject.toml
[project]
name = "databricks-openai"
version = "0.7.0.dev0"
description = "Support for Databricks AI support with OpenAI"
authors = [
    { name="Databricks", email="agent-feedback@databricks.com" },
]
readme = "README.md"
license = { text="Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "databricks-vectorsearch>=0.50",
    "langchain>=0.3.0",
    "databricks-ai-bridge>=0.4.2",
    "openai>=1.99.9",
    "pydantic>2.10.0",
    "mlflow>=2.20.1",
    "unitycatalog-openai[databricks]>=0.2.0",
    "databricks-mcp>=0.4.0",
    "openai-agents>=0.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "typing_extensions",
    "databricks-sdk>=0.34.0",
    "ruff==0.6.4",
]

integration = [
    "pytest-timeout>=2.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/databricks_openai/*"
]

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_openai"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    # isort
    "I",
    # bugbear rules
    "B",
    # remove unused imports
    "F401",
    # bare except statements
    "E722",
    # print statements
    "T201",
    "T203",
    # misuse of typing.TYPE_CHECKING
    "TCH004",
    # import rules
    "TID251",
    # undefined-local-with-import-star
    "F403",
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.ruff.lint.pydocstyle]
convention = "google"
